move _try_command to XiaomiMiioEntity class
